+++ src/feature/search/SearchSuggestionResult.tsx
