+++ src/feature/simulation/SimulateResultCard.tsx
