+++ src/feature/simulation/SimulationSliderControl.tsx
